backend/requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/dev.in
#
asgiref==3.5.2
    # via django
autopep8==1.6.0
    # via
    #   -r requirements/dev.in
    #   django-silk
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
coverage==6.3.3
    # via -r requirements/dev.in
cryptography==37.0.2
    # via django-rest-knox
django==3.2.13
    # via
    #   -r requirements/base.in
    #   django-cors-headers
    #   django-filter
    #   django-guardian
    #   django-rest-knox
    #   django-silk
    #   django-typomatic
    #   djangorestframework
    #   drf-nested-routers
    #   drf-yasg
django-cors-headers==3.12.0
    # via -r requirements/base.in
django-environ==0.8.1
    # via -r requirements/base.in
django-filter==21.1
    # via -r requirements/base.in
django-guardian==2.4.0
    # via -r requirements/base.in
django-rest-knox==4.2.0
    # via -r requirements/base.in
django-silk==4.3.0
    # via -r requirements/dev.in
django-typomatic==1.6.1
    # via -r requirements/base.in
djangorestframework==3.13.1
    # via
    #   -r requirements/base.in
    #   django-rest-knox
    #   django-typomatic
    #   drf-nested-routers
    #   drf-yasg
drf-nested-routers==0.93.4
    # via -r requirements/base.in
drf-yasg==1.20.0
    # via -r requirements/dev.in
factory-boy==3.2.1
    # via -r requirements/dev.in
faker==13.11.1
    # via factory-boy
flake8==4.0.1
    # via -r requirements/dev.in
gprof2dot==2021.2.21
    # via django-silk
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via
    #   coreschema
    #   django-silk
markupsafe==2.1.1
    # via jinja2
mccabe==0.6.1
    # via flake8
packaging==21.3
    # via drf-yasg
psycopg2==2.9.3
    # via -r requirements/base.in
pycodestyle==2.8.0
    # via
    #   autopep8
    #   flake8
pycparser==2.21
    # via cffi
pyflakes==2.4.0
    # via flake8
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via
    #   django-silk
    #   faker
pytz==2022.1
    # via
    #   django
    #   django-silk
    #   djangorestframework
pyyaml==6.0
    # via -r requirements/dev.in
requests==2.27.1
    # via
    #   coreapi
    #   django-silk
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via
    #   django
    #   django-silk
toml==0.10.2
    # via autopep8
uritemplate==4.1.1
    # via
    #   -r requirements/dev.in
    #   coreapi
    #   drf-yasg
urllib3==1.26.9
    # via requests
uwsgi==2.0.20
    # via -r requirements/base.in
